sdk/tests/conformance2/reading/read-pixels-pack-parameters.html

<!--

/*
** Copyright (c) 2015 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../js/js-test-pre.js"></script>
<script src="../../js/webgl-test-utils.js"></script>
</head>
<body>
<canvas id="example" width="4" height="4"></canvas>
<div id="description"></div>
<div id="console"></div>
<script>
function calculatePaddingBytes(bytesPerPixel, packAlignment, width)
{
    var padding = 0;
    switch (packAlignment) {
    case 1:
    case 2:
    case 4:
    case 8:
        padding = (bytesPerPixel * width) % packAlignment;
        if (padding > 0)
            padding = packAlignment - padding;
        return padding;
    default:
        testFailed("should not reach here");
        return;
    }
}

function runTestIteration(width, height, packParams)
{
    if (!("alignment" in packParams))
        packParams.alignment = 4;
    if (!("rowLength" in packParams))
        packParams.rowLength = 0;
    if (!("skipPixels" in packParams))
        packParams.skipPixels = 0;
    if (!("skipRows" in packParams))
        packParams.skipRows = 0;
    debug("Testing width = " + width + ", height = " + height +
          ", PACK_ALIGNMENT = " + packParams.alignment + ", PACK_ROW_LENGTH = " + packParams.rowLength +
          ", PACK_SKIP_PIXELS = " + packParams.skipPixels + " , PACK_SKIP_ROWS = " + packParams.skipRows);
    gl.pixelStorei(gl.PACK_ALIGNMENT, packParams.alignment);
    gl.pixelStorei(gl.PACK_ROW_LENGTH, packParams.rowLength);
    gl.pixelStorei(gl.PACK_SKIP_PIXELS, packParams.skipPixels);
    gl.pixelStorei(gl.PACK_SKIP_ROWS, packParams.skipRows);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR);

    gl.clearColor(expectedColor[0] / 255.0, expectedColor[1] / 255.0, expectedColor[2] / 255.0, expectedColor[3] / 255.0);
    gl.clear(gl.COLOR_BUFFER_BIT);

    var actualWidth = packParams.rowLength > 0 ? packParams.rowLength : width;

    var bytesPerPixel = 4; // see readPixels' parameters below, the format is gl.RGBA, type is gl.UNSIGNED_BYTE
    var padding = calculatePaddingBytes(bytesPerPixel, packParams.alignment, actualWidth);

    var size = bytesPerPixel * actualWidth * height + padding * (height - 1);
    if (size < 0)
        size = 0;
    var skipSize = (packParams.skipPixels * bytesPerPixel +
                    packParams.skipRows * (bytesPerPixel * actualWidth + padding));
    var buffer = new Uint8Array(size + skipSize);
    gl.readPixels(0, 0, width, height, gl.RGBA, gl.UNSIGNED_BYTE, buffer);
    if (packParams.rowLength > 0 && packParams.rowLength < width) {
        // Revisit this if spec goes the other way.
        wtu.glErrorShouldBe(gl, gl.INVALID_OPERATION);
        return;
    }
    wtu.glErrorShouldBe(gl, gl.NO_ERROR);

    if (!size)
        return;

    // Check the last pixel of the last row.
    var bytesPerRow = actualWidth * bytesPerPixel + padding;
    var pos = skipSize + bytesPerRow * (height - 1) + (width - 1) * bytesPerPixel;
    for (var i = 0; i < bytesPerPixel; ++i)
        pixel[i] = buffer[pos + i];
    shouldBe("pixel", "expectedColor");
}

function testPackParameters()
{
    debug("");
    debug("Verify that reading pixels to array buffer works fine with various pack alignments.");
    runTestIteration(1, 3, {alignment:1});
    runTestIteration(1, 3, {alignment:2});
    runTestIteration(1, 3, {alignment:4});
    runTestIteration(1, 3, {alignment:8});
    runTestIteration(2, 3, {alignment:4});
    runTestIteration(2, 3, {alignment:8});
    runTestIteration(3, 3, {alignment:4});
    runTestIteration(3, 3, {alignment:8});
    runTestIteration(0, 0, {alignment:1});
    runTestIteration(1, 3, {alignment:4});
    runTestIteration(1, 3, {alignment:8});

    debug("");
    debug("Verify that reading pixels to array buffer works fine with pack row length.");
    // PACK_ROW_LENGTH < width case, pending working group decision.
    // runTestIteration(3, 3, {alignment:8, rowLength:2});  // PACK_ROW_LENTH < width
    runTestIteration(3, 3, {alignment:8, rowLength:3});  // PACK_ROW_LENTH == width
    runTestIteration(3, 3, {alignment:8, rowLength:4});  // PACK_ROW_LENTH > width, no padding
    runTestIteration(3, 3, {alignment:8, rowLength:5});  // PACK_ROW_LENTH > width, extra padding

    debug("");
    debug("Verify that reading pixels to array buffer works fine with pack skip parameters.");
    runTestIteration(3, 3, {alignment:8, skipPixels:2});
    runTestIteration(3, 3, {alignment:8, skipPixels:1, skipRows:3});
    runTestIteration(3, 3, {alignment:8, skipRows:3});
    runTestIteration(2, 3, {alignment:8, skipPixels:2});
    runTestIteration(2, 3, {alignment:8, skipPixels:1, skipRows:3});
    runTestIteration(2, 3, {alignment:8, skipRows:3});
}

debug("");
debug("Canvas.getContext");

var wtu = WebGLTestUtils;
var pixel = [0, 0, 0, 0];
var expectedColor = [255, 102, 0, 255];

var canvas = document.getElementById("example");
var gl = wtu.create3DContext(canvas);

if (!gl) {
  testFailed("context does not exist");
} else {
  testPassed("context exists");

  debug("");
  description('ReadPixels into array buffer');
  testPackParameters();
}

debug("");
var successfullyParsed = true;
</script>
<script src="../../js/js-test-post.js"></script>
</body>
</html>
